Extract dependency query runner partial
